backend/requirements.txt
aniso8601==9.0.1
appnope==0.1.4
asgiref==3.8.1
asttokens==3.0.0
attrs==24.3.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.2.0
Brotli==1.1.0
certifi==2024.12.14
charset-normalizer==3.4.1
decorator==5.1.1
defusedxml==0.7.1
dj-database-url==2.3.0
Django==5.1.7
django-admin-sortable2==2.2.4
django-cleanup==8.1.0
django-compression-middleware==0.5.0
django-cors-headers==4.3.1
django-debug-toolbar==5.0.1
django-filter==24.2
django-graphql-jwt==0.4.0
django-reset-migrations==0.4.0
django-setty==3.2.2
django-storages~=1.14.5
django-stubs==4.2.7
django-stubs-ext==5.0.2
django-tinymce==4.0.0
django-versatileimagefield==3.1
django-vite==3.0.4
docopt==0.6.2
environ==1.0
executing==2.1.0
fastjsonschema==2.21.1
graphene==3.3
graphene-django==3.2.1
graphql-core==3.2.3
graphql-relay==3.2.0
idna==3.10
ipython==8.12.3
jedi==0.19.2
Jinja2==3.1.5
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mistune==3.1.0
mypy==1.7.1
mypy-extensions==1.0.0
nbclient==0.10.2
nbconvert==7.16.5
nbformat==5.10.4
packaging==24.2
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==11.1.0
pipreqs==0.4.13
platformdirs==4.3.6
promise==2.3
prompt_toolkit==3.0.48
psycopg2-binary==2.9.10; platform_system == "Windows"
psycopg2==2.9.9; platform_system == "Linux"
psycopg2-binary==2.9.10; platform_system == "darwin"
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.1
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-magic==0.4.27
python-magic-bin==0.4.14; platform_system == "Windows"
pytz==2024.1
pyzmq==26.2.0
referencing==0.35.1
requests==2.32.3
rpds-py==0.22.3
six==1.16.0
soupsieve==2.6
sqlparse==0.5.0
setuptools==76.0.0
stack-data==0.6.3
strawberry-django-auth==0.378.4
strawberry-graphql==0.260.4
strawberry-graphql-django==0.56.0
text-unidecode==1.3
tinycss2==1.4.0
tornado==6.4.2
traitlets==5.14.3
types-pytz==2024.1.0.20240417
types-PyYAML==6.0.12.20240311
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.3.0
wcwidth==0.2.13
webencodings==0.5.1
whitenoise==6.6.0
yarg==0.1.9
zstandard==0.23.0
zstd~=1.5.6.2